redefining the primary yakonfig API:
  set_global_config(arg1)

now expects just one argument, and it can be either a path to a yaml
file, an stream object open for reading to get yaml, or a mapping
object.  This allows a program to pass its config as a dict to a child
process without having to tell the child process to reload from a file
or something complicated.

Introducing clear_global_config as a public API for resetting the
config.  Currently, the only use case for this is tests in packages
using yakonfig.