Submodules, set_default_config extensions.
A Configurable declaration can declare sub_modules, which is a list
of Configurable objects, and their configs should be contained in this
object's config.

set_default_config() can be passed a YAML string or filename which
will be overlaid as though it were the input config.
"with defaulted_config(...)" resets the global config after completion
(and is really for tests).