Remove support for !runtime in configuration.
This was an older way to support command-line arguments where a runtime
dictionary could be pushed into yakonfig, and "!runtime key" as a value
would get values from it.  The current merging of default values,
user configuration, and command-line options replaces this, and all code
seems to have migrated to the newer version.