Support Python 2.6; remove !include_func.
yakonfig previously supported syntax where the result of a Python function
call could be inserted into the generated configuration.  This is a bad
idea for several reasons, and to my knowledge it's unused.  If you really
need it, yakonfig doesn't use "safe" PyYAML mode, so you can use
!!python/object/apply:module.function [param, param, ...]
to the same effect.

This used the importlib module, which was new in Python 2.7.  Removing this
makes all tests pass on Python 2.6.